Turn off the display 10s after playback stops
The screensaver does not automatically turn off the display after
being enabled; it only does so 10s after an input event on the screen.
So, previous behaviour was that the screen would typically be left
on, because input events (i.e. a touch of the UI) are relatively rare.
This change switches off the display 10s after the transition from
playing to not playing, as a one off.